+++ src/ECS/Systems/CameraSystem.h
@@ -1,0 +1,7 @@
+#pragma once
+#include <entt/entt.hpp>
+
+class CameraSystem {
+public:
+    void Update(entt::registry& registry, float realDt);
+};