+++ src/ECS/Systems/ProductionSystem.h
@@ -1,0 +1,7 @@
+#pragma once
+#include <entt/entt.hpp>
+
+class ProductionSystem {
+public:
+    void Update(entt::registry& registry, float realDt);
+};